Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -6,7 +6,7 @@ char* welcome(num) {
 }
 
 int main() {
-    double float num;
+    float num;
     welcome(num);
     return 0;
 }
